switched to floats for more accuracy

+++ db/migrate/20151018202238_add_weighted_score_to_convos.rb
@@ -1,5 +1,5 @@
 class AddWeightedScoreToConvos < ActiveRecord::Migration
   def change
-    add_column :convos, :weighted_score, :integer, default: 0
+    add_column :convos, :weighted_score, :float, default: 0
   end
 end
